refactor: move class serialization to data_class module.
